Renamed challengeModel.js to models.js
